Create FindTermInMacros
Allows you to provide a string search term which is then compared to the definitions of all macros (embedded and stand alone) to see if the term appears in the definition. Includes code that converts all stand alone macros to text and saves them to a file. Can also be modified to allow you to search through all forms/report/controls for anything you need.

Source: https://www.devhut.net/2017/05/25/ms-access-find-macros-using-a-search-term/