Create MultipleClickableSortableColumnsOnContinuousFormsFunctions
The following library of code can be placed in a module that is shared among all application objects. It provides all functionality needed to make it so that a continuous form (a form that displays multiple records at once without being in datasheet view), can have column headers that when clicked on, will adjust the sort order of the records on the form.