Improved instructions
Added 2 steps: 5) includes code for a dependent function that was in my GlobalFunctions modules so I missed its dependency and 6) includes instructions for copying the ArrayFunctions into a module in your project. Also changed existing steps that asked the user to copy/paste code into separate modules to simply allow them to uncomment the lines to include them in the current module, but also give them instructions for how to copy/paste into separate modules if they wanted.